reorganize files so that modules are more self-contained
