+++ src/modules/organizations/containers/organizations-container.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import apiClient from 'panoptes-client/lib/api-client';
 
-import { setOrganizations } from '../../common/actions/organizations';
+import { setOrganizations } from '../actionCreators';
 import { organizationShape, organizationsShape } from '../model';
 import OrganizationsLayout from '../components/organizations-layout';
