+++ src/modules/organizations/reducer.js
@@ -1,5 +1,5 @@
-import * as types from '../constants/action-types';
-import initialState from './initial-state';
+import * as types from '../../constants/action-types';
+import initialState from '../../initial-state';
 
 export default function (state = initialState, action) { // eslint-disable-line import/prefer-default-export
   switch (action.type) {
