+++ src/modules/users/actionCreators.js
@@ -1,0 +1,8 @@
+import * as types from '../../constants/action-types';
+
+export function setLoginUser(user) { // eslint-disable-line import/prefer-default-export
+  return {
+    type: types.SET_LOGIN_USER,
+    user,
+  };
+}
